[nemo-qml-plugin-dbus] Add an auto quit feature for DBus adaptors. Contributes to JB#54789
Add an option to a acquire an event loop locker at startup and release
it after a timeout. This will close the application if no windows are
visible at that time.

And allow additional interfaces to defined for an object by grouping
multiple adaptors into an object for a path. For applications which
need to support multiple interfaces but can group them on a common path
this allows them to continue letting an object register the service
without conflicts or race conditions between the adaptors. Applications
which have to support multiple objects will also have to register
themselves.